rename directory and extract components
